backend e2e tests

+++ apps/student-calendar-backend/src/app/tasks/tasks.service.ts
@@ -22,12 +22,36 @@ export class TasksService {
     private userService: UsersService
   ) {}
 
-  async getByClass(classId: string, afterDate?: Date): Promise<Task[]> {
+  async getByClass(
+    classId: string,
+    userId: string,
+    afterDate?: Date
+  ): Promise<Task[]> {
+    const currentClass = await this.classesService.getById(classId);
+    if (!currentClass) throw new NotFoundException('Class does not exist');
+    const isStudentInClass = currentClass.students.some(
+      (el) => el.userId === userId
+    );
+    if (!isStudentInClass)
+      throw new UnauthorizedException('Student is not in class');
     return this.repo.getTasksForClass(classId, afterDate);
   }
 
-  async getById(id: string): Promise<Task & { student: Student }> {
-    return this.repo.getById(id);
+  async getById(
+    id: string,
+    classId: string,
+    userId: string
+  ): Promise<Task & { student: Student }> {
+    const currentClass = await this.classesService.getById(classId);
+    if (!currentClass) throw new NotFoundException('Class does not exist');
+    const isStudentInClass = currentClass.students.some(
+      (el) => el.userId === userId
+    );
+    if (!isStudentInClass)
+      throw new UnauthorizedException('Student is not in class');
+    const task = await this.repo.getById(id, classId);
+    if (!task) throw new NotFoundException('Task not found');
+    return task;
   }
 
   async add(
@@ -44,30 +68,33 @@ export class TasksService {
       userId,
       classId
     );
-    const isStudentInClass = currentClass.students.some(
-      (el) => el.id === student.id
-    );
-    if (!isStudentInClass)
+    if (!student)
       throw new UnauthorizedException('Student not in class');
     const result = await this.repo.create({
       ...newTask,
       classId,
       createdBy: student.id,
     });
-    const notificationDate = startOfHour(setHours(result.deliverDate, 8));
+    if (process.env.NODE_ENV !== 'test') {
 
-    if (differenceInHours(notificationDate, new Date()) >= MIN_TIME_FOR_NOTIFICATION) {
-      await this.notificationService.scheduleNotification(
-        `/classes/${classId}/tasks/${result.id}/notify`,
-        {
-          title: 'Task coming up',
-          body: `The task ${result.name} is scheduled to today`,
-          data: JSON.stringify({
-            taskId: result.id,
-          }),
-        },
-        notificationDate
-      );
+      const notificationDate = startOfHour(setHours(result.deliverDate, 8));
+
+      if (
+        differenceInHours(notificationDate, new Date()) >=
+        MIN_TIME_FOR_NOTIFICATION
+      ) {
+        await this.notificationService.scheduleNotification(
+          `/classes/${classId}/tasks/${result.id}/notify`,
+          {
+            title: 'Task coming up',
+            body: `The task ${result.name} is scheduled to today`,
+            data: JSON.stringify({
+              taskId: result.id,
+            }),
+          },
+          notificationDate
+        );
+      }
     }
     return result;
   }
@@ -80,10 +107,10 @@ export class TasksService {
       Omit<Task, 'id' | 'createdBy' | 'classId' | 'createdAt' | 'updatedAt'>
     >
   ) {
-    const currentTask = await this.getById(id);
-    if (!currentTask) throw new NotFoundException('Task not found');
     const currentClass = await this.classesService.getById(classId);
     if (!currentClass) throw new NotFoundException('Class not found');
+    const currentTask = await this.getById(id, classId, userId);
+    if (!currentTask) throw new NotFoundException('Task not found');
     const student = await this.studentService.getStudentByUserIdAndClassId(
       userId,
       classId
@@ -103,7 +130,8 @@ export class TasksService {
       userId,
       classId
     );
-    const currentTask = await this.getById(id);
+    if (!student) throw new UnauthorizedException('Student is not in class');
+    const currentTask = await this.getById(id, classId, userId);
     if (currentTask.createdBy !== student.id)
       throw new UnauthorizedException('User cannot delete this task');
     return this.repo.deleteById(id, student.id);
@@ -115,7 +143,10 @@ export class TasksService {
       currentClass.students.map(async ({ userId }) => {
         const user = await this.userService.getByUserId(userId);
         if (user && user.fcmToken)
-          return this.notificationService.sendNotification(user.fcmToken, payload);
+          return this.notificationService.sendNotification(
+            user.fcmToken,
+            payload
+          );
       })
     );
   }
